Improvements to support 2.1 RC1 and no compilation
This commits introduces a flag to deploy on dotcloud without
recompilation. This is interesting because compilation is the step which
usually is the memory greediest. As play 2.1 now only loads
only direct dependencies of your project, it means than you can deploy
great typesafe web applications with a very limited memory consumption.

Less memory, less money[1] !

1: https://www.dotcloud.com/pricing.html